requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
sentence-transformers
torch
google-generativeai
